+++ src/apps/obsidian-plugin/components/Select/useAccountSelect.tsx
@@ -1,36 +1,25 @@
 import { AccountsContext } from "apps/obsidian-plugin/views/RightSidebarReactView/Contexts";
-import { Account, AccountName } from "contexts/Accounts/domain";
-import { useContext, useEffect, useMemo, useState } from "react";
+import { Account } from "contexts/Accounts/domain";
+import { useContext, useEffect, useState } from "react";
 import { Nanoid } from "../../../../contexts/Shared/domain";
 import { Select } from "./Select";
 
 export const useAccountSelect = ({
 	label,
-	initialValueName,
 	initialValueID,
 	error,
 }: {
 	label?: string;
-	initialValueName?: AccountName;
 	initialValueID?: string;
 	error?: string;
 }) => {
-	const [accountName, setAccountName] = useState(
-		initialValueName?.value ?? "",
-	);
+	const [accountId, setAccountId] = useState(initialValueID ?? "");
 	const [account, setAccount] = useState<Account>();
 
 	const { accounts, getAccountByID } = useContext(AccountsContext);
-	const accountNames = useMemo(
-		() =>
-			accounts
-				.map((acc) => acc.name.value)
-				.toSorted((a, b) => a.localeCompare(b)),
-		[accounts],
-	);
 
 	useEffect(() => {
-		setAccountName(
+		setAccountId(
 			initialValueID
 				? (getAccountByID(new Nanoid(initialValueID))?.name.value ?? "")
 				: "",
@@ -39,22 +28,22 @@ export const useAccountSelect = ({
 
 	useEffect(() => {
 		setAccount(
-			accountName
-				? accounts.find((acc) =>
-						acc.name.equalTo(new AccountName(accountName)),
-					)
+			accountId
+				? accounts.find((acc) => acc.id.equalTo(new Nanoid(accountId)))
 				: undefined,
 		);
-	}, [accountName]);
+	}, [accountId]);
 
 	return {
 		AccountSelect: (
-			<Select
+			<Select<Account>
 				id="account"
 				label={label ?? "Account"}
-				value={accountName}
-				values={["", ...accountNames]}
-				onChange={(account) => setAccountName(account)}
+				value={accountId}
+				values={accounts}
+				getOptionLabel={(account) => account.name.value}
+				getOptionValue={(account) => account.id.value}
+				onChange={(account) => setAccountId(account)}
 				error={error}
 			/>
 		),